requirements.txt
discord
python-dotenv
yt_dlp
asyncio
pynacl
youtube-search-python